refactor(api/applications): improve code and structure in exam timetable
