+++ apps/api/src/server/applications/examination-timetable-items/examination-timetable-items.service.js
@@ -2,11 +2,16 @@ import { eventClient } from '#infrastructure/event-client.js';
 import { NSIP_EXAM_TIMETABLE } from '#infrastructure/topics.js';
 import * as examinationTimetableRepository from '../../repositories/examination-timetable.repository.js';
 import * as examinationTimetableItemsRepository from '../../repositories/examination-timetable-items.repository.js';
+import * as examinationTimetableTypesRepository from '../../repositories/examination-timetable-types.repository.js';
+import * as documentRepository from '../../repositories/document.repository.js';
+import * as folderRepository from '../../repositories/folder.repository.js';
+import logger from '../../utils/logger.js';
 import { EventType } from '@pins/event-client';
 
 /**
  * @typedef {import('../../../message-schemas/events/nsip-exam-timetable.js').NSIPExamTimetableItem} NSIPExamTimetableItem
  * @typedef {import('../../../message-schemas/events/nsip-exam-timetable.js').NSIPExamTimetableItemDescriptionLineItem} NSIPExamTimetableItemDescriptionLineItem
+ * @typedef {import('@pins/applications.api').Schema.Folder} Folder
  */
 
 /**
@@ -53,7 +58,7 @@ async function buildExamTimetableItemsPayload(examinationTimetableId) {
 		return [];
 	}
 
-	return examinationTimetableItems?.map(buildSingleExaminationTimetableItemPayload);
+	return examinationTimetableItems?.map(buildSingleExaminationTimetableItemPayload) ?? [];
 }
 
 /**
@@ -110,3 +115,125 @@ export async function unPublish(id) {
 		published: false
 	});
 }
+
+/**
+ *
+ * @param {import('@pins/applications.api').Schema.ExaminationTimetableItem} examinationTimetableItem
+ * @param {Number} parentFolderId
+ * @param {Number} caseId
+ * @returns
+ */
+export const createDeadlineSubFolders = async (
+	examinationTimetableItem,
+	parentFolderId,
+	caseId
+) => {
+	if (!examinationTimetableItem?.description) {
+		return;
+	}
+
+	const description = JSON.parse(examinationTimetableItem?.description);
+	const categoryType = await examinationTimetableTypesRepository.getById(
+		examinationTimetableItem.examinationTypeId
+	);
+
+	if (
+		!categoryType?.name ||
+		(categoryType?.name !== 'Deadline' &&
+			categoryType?.name !== 'Procedural Deadline (Pre-Examination)')
+	) {
+		logger.info('Category name does not match to Deadline, skip creating sub folders');
+		return;
+	}
+
+	/**
+	 * @type {Promise<Folder |null>[]}
+	 */
+	const createFolderPromise = [];
+
+	// create Other sub folder
+	const otherFolder = {
+		displayNameEn: 'Other',
+		caseId,
+		parentFolderId: parentFolderId,
+		displayOrder: 100
+	};
+	createFolderPromise.push(folderRepository.createFolder(otherFolder));
+
+	if (!description?.bulletPoints || description?.bulletPoints?.length === 0) {
+		logger.info('No bulletpoints');
+	}
+
+	// create sub folder for each bullet points.
+	description.bulletPoints.forEach((/** @type {String} */ folderName) => {
+		const subFolder = {
+			displayNameEn: folderName?.trim(),
+			caseId,
+			parentFolderId: parentFolderId,
+			displayOrder: 100
+		};
+		createFolderPromise.push(folderRepository.createFolder(subFolder));
+	});
+
+	logger.info('Create sub folders');
+	await Promise.all(createFolderPromise);
+	logger.info('Sub folders created successfully');
+};
+
+/**
+ * Deletes all the sub folders in a exam timetable folder, assumption is that all folders are empty.
+ * eg for deadline exams, deletes all line item folders and the "Other" folder
+ *
+ * @param {Number} caseId
+ * @param {Number} parentFolderId
+ * @returns {import('@prisma/client').PrismaPromise<import('@pins/applications.api').Schema.BatchPayload>}
+ */
+export const deleteDeadlineSubFolders = async (caseId, parentFolderId) => {
+	const subFolders = await folderRepository.getByCaseId(caseId, parentFolderId);
+	if (!subFolders) {
+		logger.info(`No sub folder found for the parent folder Id ${parentFolderId}`);
+		return;
+	}
+
+	const idsToDelete = subFolders.map(folder => folder.id);
+	await folderRepository.deleteFolderMany(idsToDelete);
+	logger.info(`Sub folders deleted successfully in folder: ${parentFolderId}`);
+};
+
+/**
+ *
+ * @param {import('@prisma/client').ExaminationTimetableItem} timetableItem
+ * @param {number} caseId
+ * @returns {Promise<boolean>}
+ */
+export const validateSubmissions = async (timetableItem, caseId) => {
+	const folder = await folderRepository.getById(timetableItem.folderId);
+	if (!folder) {
+		logger.info('No associated folder found');
+		return false;
+	}
+
+	const documents = await documentRepository.countDocumentsInFolder(timetableItem.folderId);
+	if (documents && documents > 0) {
+		logger.info(`Document found in folder ${timetableItem.folderId}`);
+		return true;
+	}
+
+	const subFolders = await folderRepository.getByCaseId(caseId, timetableItem.folderId);
+	if (!subFolders || subFolders.length === 0) {
+		logger.info(`No sub folder for folder ${timetableItem.folderId}`);
+		return false;
+	}
+
+	const subfoldersDocuments = await Promise.all(
+		subFolders.map((subFolder) => documentRepository.countDocumentsInFolder(subFolder.id))
+	);
+
+	if (subfoldersDocuments.some((sd) => sd > 0)) {
+		logger.info(`Document found for parent folder ${timetableItem.folderId}`);
+		return true;
+	}
+
+	logger.info(`Document not found for parent folder ${timetableItem.folderId}`);
+	return false;
+};
